Make listeners.js file
